requirements.txt
flask==2.0.1
flask-cors==3.0.10
google-generativeai==0.1.0
Pillow==10.2.0
python-dotenv==0.19.0
serverless-wsgi==3.0.1
gunicorn==20.1.0
